Add DrawSprite routine
Took forever to figure out why it was segfaulting, only to discover it was because GIMP was saving only RGB (24-bit) data instead of RGBA (32-bit).

Also, SDL seems to have the surfaces in BGRA pixel format instead of RGBA mode, so I had to swap B&R channels in the original picture for it to look proper in here.  Need to figure out how to make SDL use RGBA since that's the format image editors assume (no "BGRA" option for "RGB Save Type" in GIMP's export options for raw images)